Minimally support namespacing of Test Mocks
- mirrors the current behavior of wrapping schema
types in a caseless enum whose name matches the
schemaName when the SchemaTypesFileOutput is
embeddedInTarget and the OperationsFileOutput is
inSchemaModule.
- when the above conditions are met, and when
TestMockFileOutput is absolute, test mock types
will be wrapped in a caseless enum whose name
matches the schemaName with the suffix "TestMocks"
(mirroring the naming convention for generated
test mocks modules).
- adds new MockNamespaceFileGenerator and
MockNamespaceTemplate
- updates MockObjectTemplate, MockUnionsTemplate,
and MockInterfacesTemplate to use detachedTemplate
for their content which should not be wrapped in
a namespace
- updates TemplateRenderer/renderTestMockFile() to
check for and include the detachedTemplate
- includes a new convenience computed var
needsTypesWrappedInNamespace on
ApolloCodegenConfiguration which consolidates the
internal logic for determining whether namespacing
should be used when rendering the relevant files
from template.
- includes a test mirroring the same validation
logic used for the conditionallyWrapInNamespace
test on operation files, as well as new
MockNamespaceFileGeneratorTests mirroring the
tests from SchemaModuleFileGeneratorTests
- includes a new MockObjectTemplate test mirroring
generatesConvenienceInitializer for when types
must be fully qualified under namespacing
- updates other tests where relevant