Minimally support namespacing of Test Mocks
- mirrors the current behavior of wrapping schema
types in a caseless enum whose name matches the
schemaName when the SchemaTypesFileOutput is
embeddedInTarget and the OperationsFileOutput is
inSchemaModule.
- when the above conditions are met, and when
TestMockFileOutput is absolute, test mock types
will be wrapped in a caseless enum whose name
matches the schemaName with the suffix "TestMocks"
(mirroring the naming convention for generated
test mocks modules).
- includes a test mirroring the same validation
logic used for the conditionallyWrapInNamespace
test on operation files